ci: fix the failing job, and clear ruff across tools/ and benchmark/
Install step failed because 'uv pip install --system' had no system Python to
target -- 'uv python install' provides a uv-managed interpreter that --system
cannot see. Added actions/setup-python before uv, bumped setup-uv to v5 with
caching.

Also cleared 31 ruff findings that 'ruff check .' surfaces outside src/tests
(the video tooling and benchmark harness). Two of them were in embedded JS
inside record_termcast.py, where an appended '# noqa' lands *inside* the
JavaScript string and would break the renderer -- those use a per-file-ignore
instead, since physical line length is meaningless in an embedded script.

ruff clean, mypy strict clean (32 files), 131 tests passing.

+++ tools/video/sketch/_style.py
@@ -36,10 +36,10 @@ import matplotlib
 
 matplotlib.use("Agg")
 
-import matplotlib.pyplot as plt  # noqa: E402
-from matplotlib import patheffects as pe  # noqa: E402
-from matplotlib.lines import Line2D  # noqa: E402
-from matplotlib.patches import (  # noqa: E402
+import matplotlib.pyplot as plt
+from matplotlib import patheffects as pe
+from matplotlib.lines import Line2D
+from matplotlib.patches import (
     Arc,
     Ellipse,
     FancyArrowPatch,
@@ -266,7 +266,7 @@ def render_sequence(state, draw, duration: float, out: Path, frames_dir: Path,
     for stale in frames_dir.glob("f*.png"):
         stale.unlink()
 
-    n = int(round(duration * fps))
+    n = round(duration * fps)
     prev_key = None
     prev_path: Path | None = None
     rendered = 0
